# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ShopFlow ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersPage"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Invoices
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentsPage"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Employees
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [productId, productName, quantity, unitPrice, total]
      properties:
        productId:   { type: integer }
        productName: { type: string }
        sku:         { type: string }
        quantity:    { type: integer }
        unitPrice:   { type: number }
        discount:    { type: number }
        gstPercent:  { type: number }
        total:       { type: number }
    Order:
      type: object
      required: [id, orderNumber, customerId, customerName, type, status, subtotal, discount, gstAmount, total, paymentStatus, createdAt]
      properties:
        id:            { type: integer }
        orderNumber:   { type: string }
        customerId:    { type: integer }
        customerName:  { type: string }
        type:          { type: string }
        status:        { type: string }
        subtotal:      { type: number }
        discount:      { type: number }
        gstAmount:     { type: number }
        total:         { type: number }
        paidAmount:    { type: number }
        paymentStatus: { type: string }
        paymentMethod: { type: ["string", "null"] }
        notes:         { type: ["string", "null"] }
        items:         { type: array, items: { $ref: "#/components/schemas/OrderItem" } }
        createdAt:     { type: string }
    OrdersPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:  { type: array, items: { $ref: "#/components/schemas/Order" } }
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }
    OrderItemInput:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:  { type: integer }
        quantity:   { type: integer }
        unitPrice:  { type: number }
        discount:   { type: number }
        gstPercent: { type: number }
    OrderInput:
      type: object
      required: [customerId, type, items]
      properties:
        customerId:    { type: integer }
        type:          { type: string, enum: [retail, wholesale] }
        discount:      { type: number }
        paymentMethod: { type: string }
        paidAmount:    { type: number }
        notes:         { type: string }
        items:         { type: array, items: { $ref: "#/components/schemas/OrderItemInput" } }
    OrderUpdate:
      type: object
      properties:
        status:        { type: string }
        paymentStatus: { type: string }
        paidAmount:    { type: number }
        paymentMethod: { type: string }
        notes:         { type: string }
...
    Payment:
      type: object
      required: [id, referenceNumber, amount, method, type, entityType, entityId, entityName, createdAt]
      properties:
        id:              { type: integer }
        referenceNumber: { type: string }
        amount:          { type: number }
        method:          { type: string }
        type:            { type: string }
        entityType:      { type: string }
        entityId:        { type: integer }
        entityName:      { type: string }
        notes:           { type: ["string", "null"] }
        createdAt:       { type: string }
    PaymentsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:  { type: array, items: { $ref: "#/components/schemas/Payment" } }
        total: { type: integer }
        page:  { type: integer }
        limit: { type: integer }
    PaymentInput:
      type: object
      required: [amount, method, type, entityType, entityId]
      properties:
        amount:     { type: number }
        method:     { type: string, enum: [cash, upi, card, bank, cheque, credit] }
        type:       { type: string, enum: [received, paid, refund] }
        entityType: { type: string, enum: [customer, supplier] }
        entityId:   { type: integer }
        notes:      { type: string }
...